feat: separate SearchItem from SearchSelect
