+++ packages/web/package.json
@@ -28,6 +28,7 @@
     "immer": "^10.0.4",
     "loglevel": "^1.9.1",
     "next": "14.1.3",
+    "prop-types": "^15.8.1",
     "react": "^18",
     "react-dom": "^18",
     "styled-components": "^6.1.8",
